Move validation of plugins
